+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [2.5.0-0] - 2025-09-10
+
 - New itertool: `groupBy`
 - Renamed `groupby` to `igroupby`
 - Deprecated `groupby` (now an alias to `igroupby`)